Explicitly require `shellesc` package
As discussed privatly at Matrix.org, `\write18` working in LuaTeX is a happy accident. In LuaTeX, the `\write18` primitive has been deprecated and removed in favor of the `os.execute()` Lua built-in. The [shellesc LaTeX package][1] polyfills `\write18`, so that it works even in LuaTeX. Shellesc is being loaded transitively by other LaTeX packages, which means that it could cease being loaded at any point in the future, breaking our code. To prevent future breakage, we should make our reliance on shellesc known. This pull request explicitly loads shellesc and denotes why we need it in a comment.

 [1]: https://www.ctan.org/pkg/shellesc